LLD/pyproject.toml
[project]
name = "streamlit-app-for-lld"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.48.1",
    "openai>=1.23.0",
    "python-dotenv>=1.0.0",
]
